+++ src/components/feed-sources/templates/calendar-api-feed-type.jsx
@@ -19,7 +19,12 @@ const CalendarApiFeedType = ({
   useEffect(() => {
     if (feedSourceId && feedSourceId !== "") {
       ConfigLoader.loadConfig().then((config) => {
-        setOptionsEndpoint(`${config.api + feedSourceId}/config/locations`);
+        let endpoint = config.api;
+        endpoint = endpoint.replace(/\/$/, "");
+        endpoint += feedSourceId;
+        endpoint += "/config/locations";
+
+        setOptionsEndpoint(endpoint);
       });
     }
   }, [feedSourceId]);
@@ -35,6 +40,7 @@ const CalendarApiFeedType = ({
         <MultiselectFromEndpoint
           onChange={handleInput}
           name="locations"
+          disableSearch={false}
           label={t("locations")}
           value={formStateObject.locations ?? []}
           optionsEndpoint={optionsEndpoint}